- 库存服务的字段类型没有对应上。。

+++ seata-order-service2002/src/main/java/com/zq/springcloud/dao/StorageDao.java
@@ -13,6 +13,6 @@ import org.apache.ibatis.annotations.Param;
 public interface StorageDao {
 
     //扣减库存信息
-    void decrease(@Param("productId") Long productId, @Param("count") Integer count);
+    void decrease(@Param("productId") Long productId, @Param("count") Long count);
 
 }
