+++ seata-order-service2002/src/main/java/com/zq/springcloud/service/StorageService.java
@@ -8,5 +8,5 @@ package com.zq.springcloud.service;
  */
 public interface StorageService {
     // 扣减库存
-    void decrease(Long productId, Integer count);
+    void decrease(Long productId, Long count);
 }
