+++ seata-order-service2002/src/main/java/com/zq/springcloud/service/impl/StorageServiceImpl.java
@@ -24,7 +24,7 @@ public class StorageServiceImpl implements StorageService {
 
     // 扣减库存
     @Override
-    public void decrease(Long productId, Integer count) {
+    public void decrease(Long productId, Long count) {
         LOGGER.info("------->storage-service中扣减库存开始");
         storageDao.decrease(productId,count);
         LOGGER.info("------->storage-service中扣减库存结束");
